requirements.txt
Django==3.0.4
djangorestframework
markdown
django-filter